+++ R/CFtime-package.R
@@ -10,7 +10,10 @@
 #' POSIXt). The CF time coordinate is formally defined in the
 #' [CF Metadata Conventions document](https://cfconventions.org/Data/cf-conventions/cf-conventions-1.12/cf-conventions.html#time-coordinate).
 #'
-#' The package can create a [CFTime] instance from scratch or, more commonly, it
+#' This package also supports the creation of a "time" dimension, using class
+#' `CFClimatology`, for climatological statistics as defined [here](https://cfconventions.org/Data/cf-conventions/cf-conventions-1.12/cf-conventions.html#climatological-statistics).
+#'
+#' The package can create a [CFTime] or [CFClimatology] instance from scratch or, more commonly, it
 #' can use the dimension attributes and dimension variable values from a netCDF
 #' resource. The package does not actually do any of the reading and the user is
 #' free to use their netCDF package of preference. The recommended package to