+++ NEWS.md
@@ -3,6 +3,9 @@
 * New class `CFClimatology` to manage climatological statistics. Method
 `CFTime$factor()` with an `era` argument will now have a `CFClimatology`
 instance attached to the attribute "CFTime" of its return value.
+* New class `CFCalendarNone` for a calendar with no annual cycle. Such a 
+calendar represents a single instant in time, which may be useful for 
+experiments on a fixed moment during the year.
 * Fixed new `CFTime` instance for factor with a "day" or "dekad" period where
 the last factor level is not at the end of the year.
 * `CFTime$new()` with a single character string offset will generate only the